// tina/types.ts
export declare type StringField = (
  | FieldGeneric<string, undefined>
  | FieldGeneric<string, true>
  | FieldGeneric<string, false>
) &
  BaseField & {
    type: "string";
    isTitle?: boolean;
    isBody?: boolean;
    options?: Option[];
  };
export declare type NumberField = (
  | FieldGeneric<number, undefined>
  | FieldGeneric<number, true>
  | FieldGeneric<number, false>
) &
  BaseField & {
    type: "number";
  };
export declare type BooleanField = (
  | FieldGeneric<boolean, undefined>
  | FieldGeneric<boolean, true>
  | FieldGeneric<boolean, false>
) &
  BaseField & {
    type: "boolean";
  };
declare type DateFormatProps = {
  /**
   * Customize the way the format is rendered
   * ```
   * dateFormat: 'YYYY MM DD'
   * ```
   */
  dateFormat?: string;
  timeFormat?: string;
};
export declare type DateTimeField = (
  | FieldGeneric<string, undefined, DateFormatProps>
  | FieldGeneric<string, true, DateFormatProps>
  | FieldGeneric<string, false, DateFormatProps>
) &
  BaseField & {
    type: "datetime";
  };
export declare type ImageField = (
  | FieldGeneric<string, undefined>
  | FieldGeneric<string, true>
  | FieldGeneric<string, false>
) &
  BaseField & {
    type: "image";
  };
export declare type ReferenceField = (
  | FieldGeneric<string, undefined>
  | FieldGeneric<string, false>
) &
  BaseField & {
    type: "reference";
    /**
     * The names of the collections this field can use as a reference
     * ```ts
     * {
     *   type: 'reference',
     *   name: 'author',
     *   collections: ['author'],
     * }
     * ```
     */
    collections: string[];
  };
declare type RichTextAst = {
  type: "root";
  children: Record<string, unknown>[];
};
export declare type RichTextField<WithNamespace extends boolean = false> = (
  | FieldGeneric<RichTextAst, undefined>
  | FieldGeneric<RichTextAst, false>
) &
  BaseField & {
    type: "rich-text";
    /**
     * When using Markdown or MDX formats, this field's value
     * will be saved to the markdown body, while all other values
     * will be stored as frontmatter
     */
    isBody?: boolean;
    templates?: RichTextTemplate<WithNamespace>[];
    /**
     * By default, Tina parses markdown with MDX, this is a more strict parser
     * that allows you to use structured content inside markdown (via `templates`).
     *
     * Specify `"markdown"` if you're having problems with Tina parsing your content.
     */
    parser?:
      | {
          type: "markdown";
          /**
           * Tina will escape entities like `<` and `[` by default. You can choose to turn
           * off all escaping, or specify HTML, so `<div>` will not be turned into `\<div>`
           */
          skipEscaping?: "all" | "html" | "none";
        }
      | {
          type: "mdx";
        };
  };
// ...
export declare type ObjectField<WithNamespace extends boolean = false> = (
  | FieldGeneric<string, undefined, ObjectUiProps>
  | FieldGeneric<string, true, ObjectUiProps>
  | FieldGeneric<string, false, ObjectUiProps>
) &
  MaybeNamespace<WithNamespace> &
  BaseField &
  (
    | {
        type: "object";
        fields: Field<WithNamespace>[];
        templates?: undefined;
        ui?: Template["ui"];
      }
    | {
        type: "object";
        fields?: undefined;
        templates: Template<WithNamespace>[];
        templateKey?: string;
      }
  );
declare type Field<WithNamespace extends boolean = false> = (
  | StringField
  | NumberField
  | BooleanField
  | DateTimeField
  | ImageField
  | ReferenceField
  | RichTextField<WithNamespace>
  | ObjectField<WithNamespace>
) &
  MaybeNamespace<WithNamespace>;
export declare type TinaField<WithNamespace extends boolean = false> =
  Field<WithNamespace> & MaybeNamespace<WithNamespace>;
declare type MaybeNamespace<WithNamespace extends boolean = false> =
  WithNamespace extends true
    ? {
        namespace: string[];
      }
    : {};